# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the BuffetPOS API documentation.",
        "title": "BuffetPOS API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/customer/staff-notifications": {
            "post": {
                "description": "Notify to Staff",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Notify to Staff",
                "parameters": [
                    {
                        "description": "Notify to Staff request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.StaffNotificationRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Access Code",
                        "name": "AccessCode",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/customer/staff-notifications/{table_id}": {
            "get": {
                "description": "Get Staff Notification By Table ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Customer"
                ],
                "summary": "Get Staff Notification By Table ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Table ID",
                        "name": "table_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Access Code",
                        "name": "AccessCode",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StaffNotificationBase"
                        }
                    }
                }
            }
        },
...
        "/general/menus/best-selling": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get Best Selling Menu",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "General"
                ],
                "summary": "Get Best Selling Menu",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.NumberMenu"
                            }
                        }
                    }
                }
            }
        },
...
        "/manage/staff-notifications": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Staff Notification.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get All Staff Notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.StaffNotificationBase"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Set invoice as paid.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Update Status Staff Notification",
                "parameters": [
                    {
                        "description": "Update Staff Notification Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateStaffNotificationRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/manage/staff-notifications/{status}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get All Staff Notification By Status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Manage"
                ],
                "summary": "Get All Staff Notification By Status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Status Notification",
                        "name": "status",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Authorization",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.StaffNotificationBase"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.StaffNotificationStatus": {
            "type": "string",
            "enum": [
                "pending",
                "accepted",
                "rejected"
            ],
            "x-enum-varnames": [
                "PENDING",
                "ACCEPTED",
                "REJECTED"
            ]
        },
...
        "requests.StaffNotificationRequest": {
            "type": "object",
            "required": [
                "table_id"
            ],
            "properties": {
                "table_id": {
                    "type": "string"
                }
            }
        },
...
        "requests.UpdateStaffNotificationRequest": {
            "type": "object",
            "properties": {
                "staff_notification_id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StaffNotificationStatus"
                }
            }
        },
...
        "responses.NumberMenu": {
            "type": "object",
            "properties": {
                "categoryId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "number": {
                    "type": "integer"
                }
            }
        },
...
        "responses.StaffNotificationBase": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.StaffNotificationStatus"
                },
                "table_id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}